refactor data/components
