+++ src/app/api/v1/device/[deviceId]/status/handlers/put-status.ts
@@ -1,7 +1,7 @@
-import { updateDeviceStatus } from '@/data-access/update-device-status'
+import { updateDeviceStatus } from '@/data-access/cloudflare-d1/update-device-status'
 import { checkMacAddress } from '@/app/api/v1/device/validation/check-mac-address'
-import { getDeviceById } from '@/data-access/get-device-by-id'
-import { updateHistory } from '@/data-access/update-history'
+import { getDeviceById } from '@/data-access/cloudflare-d1/get-device-by-id'
+import { updateHistory } from '@/data-access/cloudflare-d1/update-history'
 
 export const putStatus = async (deviceId: string, status: string) => {
   checkMacAddress(deviceId)
