+++ src/data-access/cloudflare-d1/delete-device.ts
@@ -1,4 +1,4 @@
-import { runQuery } from '@/data-access/util/run-query'
+import { runQuery } from '@/data-access/cloudflare-d1/util/run-query'
 
 export const deleteDevice = async (deviceId: string, updateTime: number) => {
   return await runQuery('update device set deleteDate=? where productUid=?', [updateTime, deviceId])
