+++ src/data-access/cloudflare-d1/get-device-by-id.ts
@@ -1,4 +1,4 @@
-import { runQuery } from '@/data-access/util/run-query'
+import { runQuery } from '@/data-access/cloudflare-d1/util/run-query'
 
 export const getDeviceById = async (id: string) => {
   const d1Result =
