+++ src/data-access/cloudflare-d1/insert-device.ts
@@ -1,5 +1,5 @@
 import { DeviceResponse, NewDevice } from '@/app/api/v1/types'
-import { runQuery } from '@/data-access/util/run-query'
+import { runQuery } from '@/data-access/cloudflare-d1/util/run-query'
 
 export const insertDevice = async (device: NewDevice, updtateTime: number, status: string): Promise<DeviceResponse> => {
   const sql = 'insert into device(productUid, productType, status, lastUpdated) values (?,?,?,?)'
