+++ src/data-access/cloudflare-d1/update-history.ts
@@ -1,4 +1,4 @@
-import { runQuery } from '@/data-access/util/run-query'
+import { runQuery } from '@/data-access/cloudflare-d1/util/run-query'
 
 export const updateHistory = async (uid: string, action: string, status: string, updateTime: number) => {
   return await runQuery(
